feat: add more shelley era utxo rules check

+++ common/src/protocol_params.rs
@@ -24,7 +24,7 @@ pub struct ProtocolParams {
 impl ProtocolParams {
     /// Calculate Transaction's Mininum required fee for shelley Era
     /// Reference: https://github.com/IntersectMBO/cardano-ledger/blob/24ef1741c5e0109e4d73685a24d8e753e225656d/eras/shelley/impl/src/Cardano/Ledger/Shelley/Tx.hs#L254
-    pub fn shelley_min_fee(&self, tx_bytes: u64) -> Result<u64> {
+    pub fn shelley_min_fee(&self, tx_bytes: u32) -> Result<u64> {
         self.shelley
             .as_ref()
             .ok_or_else(|| anyhow::anyhow!("Shelley params are not set"))
@@ -146,8 +146,9 @@ pub struct ShelleyParams {
 }
 
 impl ShelleyParams {
-    pub fn min_fee(&self, tx_bytes: u64) -> u64 {
-        (tx_bytes * self.protocol_params.minfee_a as u64) + (self.protocol_params.minfee_b as u64)
+    pub fn min_fee(&self, tx_bytes: u32) -> u64 {
+        (tx_bytes as u64 * self.protocol_params.minfee_a as u64)
+            + (self.protocol_params.minfee_b as u64)
     }
 }
